Move unit tests to tests folder, change references
