+++ tempero/package.json
@@ -41,7 +41,7 @@
     "eslint-plugin-react-refresh": "^0.4.22",
     "gh-pages": "^6.3.0",
     "globals": "^16.4.0",
-    "jsdom": "^27.2.0",
+    "jsdom": "^24.0.0",
     "postcss": "^8.5.6",
     "tailwindcss": "^4.1.15",
     "typescript": "~5.9.3",
